+++ modules/tinymce/src/core/main/ts/api/Annotator.ts
@@ -33,6 +33,7 @@ interface Annotator {
   annotate: (name: string, data: DecoratorData) => void;
   annotationChanged: (name: string, f: AnnotationListenerApi) => void;
   remove: (name: string) => void;
+  removeAll: (name: string) => void;
   getAll: (name: string) => Record<string, Element[]>;
 }
 
@@ -90,6 +91,18 @@ const Annotator = (editor: Editor): Annotator => {
       identify(editor, Optional.some(name)).each(({ elements }) => {
         Arr.each(elements, Remove.unwrap);
       });
+    },
+
+    /**
+     * Removes all annotations that match the specified name from the entire document.
+     *
+     * @method removeAll
+     * @param {String} name the name of the annotation to remove
+     */
+    removeAll: (name: string): void => {
+      const bookmark = editor.selection.getBookmark();
+      Obj.each(findAll(editor, name), (spans, _) => Arr.each(spans, Remove.unwrap));
+      editor.selection.moveToBookmark(bookmark);
     },
 
     /**